유정이 리뷰

+++ SingleRior_Project/src/main/java/com/E1I4/project/common/model/vo/Review.java
@@ -26,5 +26,5 @@ public class Review {
 	private String productName;
 	private int productQuantity;
 	private String productOption;
-	private int reviewRating;
+	private double reviewRating;
 }
